Move helper functions from index.ts to appropriate libraries
